refactorizacion componente NewProduct
